+++ src/components/ComponentInterface.php
@@ -8,8 +8,8 @@ namespace VovanVE\HtmlTemplate\components;
 interface ComponentInterface
 {
     /**
-     * @param array|null $content
+     * @param \Closure|null $content
      * @return string
      */
-    public function render(?array $content = null): string;
+    public function render(?\Closure $content = null): string;
 }