+++ tests/helpers/CounterStepComponent.php
@@ -8,15 +8,15 @@ class CounterStepComponent extends BaseComponent
     public $step;
 
     /**
-     * @param array|null $content
+     * @param \Closure|null $content
      * @return string
      */
-    public function render(?array $content = null): string
+    public function render(?\Closure $content = null): string
     {
         $marker = "step: " . ($this->step ?? '?');
         if (null === $content) {
             return "<!-- $marker /-->";
         }
-        return "<!-- $marker -->" . join('', $content) . "<!-- /$marker -->";
+        return "<!-- $marker -->" . join('', $content()) . "<!-- /$marker -->";
     }
 }